refactor: mvc 구조에 맞게 패키지 분리 및 리뷰 적용
