+++ src/test/java/calculator/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package step2;
+package calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
